Changed layout and comments for Go 1.5.
